Refactor and some fixes
