+++ modin/core/dataframe/base/dataframe/protocol/test/__init__.py
@@ -10,3 +10,5 @@
 # the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific language
 # governing permissions and limitations under the License.
+
+"""Tests for the DataFrame exchange protocol."""
